extract HeroViewHolder from Hero model
